+++ changelog.md
@@ -16,6 +16,7 @@ This version is not yet released. If you are reading this on the website, then t
 - Add sided subscripts for [`negate ¯`](https://uiua.org/docs/negate), which get the conjugate or negative conjugate
 - Add numeric subscripts for [`rise ⍏`](https://uiua.org/docs/rise) and [`fall ⍖`](https://uiua.org/docs/fall), which give [`pick ⊡`](https://uiua.org/docs/pick)-compatible indices
 - Allow for multiple outputs from monadic [`tuples ⧅`](https://uiua.org/docs/tuples)
+- Allow sided [`stencil ⧈`](https://uiua.org/docs/stencil) with non-monadic functions
 - [`under ⍜`](https://uiua.org/docs/under)[`group ⊕`](https://uiua.org/docs/group) now promotes low-rank groups like [`keep ▽`](https://uiua.org/docs/keep) and others do.
 - Lone numeric subscripts now format to number literals, making it easier to split them off
 - Add `NoInit` constructor to data definitions construct data without running initializers (but still running validators)